requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.2
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
